cephadm: move nfs class to a new file
