+++ src/cephadm/cephadmlib/daemons/__init__.py
@@ -3,6 +3,7 @@ from .tracing import Tracing
 from .ingress import HAproxy, Keepalived
 from .nvmeof import CephNvmeof
 from .iscsi import CephIscsi
+from .nfs import NFSGanesha
 
 __all__ = [
     'CephIscsi',
@@ -10,5 +11,6 @@ __all__ = [
     'CustomContainer',
     'HAproxy',
     'Keepalived',
+    'NFSGanesha',
     'Tracing',
 ]
